Expanded the PRenderer interface
- PRenderer now has a method to draw an arc instead of only having a
method for drawing full ellipses. The drawEllipse method now delegates
to drawArc by default.
- PRenderer can now draw axis aligned rounded rectangles.
- The parameter list of drawEllipse has been corrected. It used int
parameters instead while all other methods used floats.
- The swing implementation of PRenderer has been updated to suppor the
new rendering methods. The LWJGL3 implementation has not yet been
updated.

+++ piq-lwjgl3/src/edu/udo/piq/lwjgl3/renderer/LwjglPRendererFbo.java
@@ -107,7 +107,7 @@ public class LwjglPRendererFbo extends LwjglPRendererBase {
 	}
 	
 	@Override
-	public void drawEllipse(int x, int y, int width, int height) {
+	public void drawEllipse(float x, float y, float width, float height) {
 		curMode.drawEllipseImmediate(x, y, width, height);
 	}
